requirements.txt
asgiref==3.9.1
cfgv==3.4.0
distlib==0.4.0
Django==5.2.6
djangorestframework==3.16.1
filelock==3.19.1
identify==2.6.14
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.34.0
